Moved common functions to common.py
