+++ permuter/vgg.py
@@ -1,4 +1,3 @@
-import copy
 from typing import Optional
 
 import numpy as np
@@ -19,8 +18,9 @@ from config import (
 )
 from helper import read_file, write_file
 from models import cifar10_loader
-from models.vgg import register_hook, train, LOOK_UP_LAYER
+from models.vgg import LOOK_UP_LAYER, register_hook, train
 from permuter._algo import ActMatching, STEstimator, WeightMatching
+from permuter.common import combine_models, get_losses
 
 WEIGHT_PERM = VGG_PERM_PATH.joinpath("weight_perm.pkl")
 ACT_PERM = VGG_PERM_PATH.joinpath("act_perm.pkl")
@@ -104,36 +104,6 @@ def permute_model(
         _col_ind = col_ind
     permuted_model.load_state_dict(perm_state_dict)
     return permuted_model
-
-
-def combine_models(
-    model1: torch.nn.Module, model2: torch.nn.Module, lam: float
-) -> torch.nn.Module:
-    """
-    Combine models using linear interpolation (1-lam)*model1 + lam*model2
-
-    :param model1: Model 1
-    :type model1: torch.nn.Module
-    :param model2: Model 2
-    :type model2: torch.nn.Module
-    :param lam: Lambda value in linear interpolation way
-    :type lam: float
-    :return: Combined model
-    :rtype: torch.nn.Module
-    """
-    # Creating dummy model
-    model3 = copy.deepcopy(model1).to(DEVICE)
-    model3_state_dict = model3.state_dict()
-    model1.to(DEVICE)
-    model2.to(DEVICE)
-    state_dict1 = model1.state_dict()
-    state_dict2 = model2.state_dict()
-
-    for key in model3_state_dict.keys():
-        model3_state_dict[key] = (1 - lam) * state_dict1[key] + lam * state_dict2[key]
-
-    model3.load_state_dict(model3_state_dict)
-    return model3
 
 
 def get_losses(
